chat_baas/requirements.txt
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.3.5
psycopg2-binary==2.7.7
py-ms==1.4.1rc2
jaeger-client==4.1.0
opentracing-instrumentation==3.2.1